pyproject.toml
[project]
name = "mcp-vllm-delegator"
version = "0.1.0"
description = "MCP server for delegating tasks to local vLLM instances"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp>=1.16.0",
    "httpx>=0.28.1",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
]
